Refactor codebase for improved organization and clarity
This commit includes significant refactoring across multiple files, focusing on the removal and addition of using directives to streamline dependencies.

Key changes include:
- Updated project references in `Program.cs` to enhance clarity.
- Modified `AddProject` method calls for correct type references.
- Adjusted using directives in `EfSampleRepository.cs`, `MigrationWorker.cs`, and migration files for consistency.
- Introduced nullable reference types in migration files for better null handling.
- Updated validation rules in `Program.cs` to use more generic types.
- Refactored repository interfaces and implementations for better abstraction.
- Ensured correct service registration in `ServiceCollectionExtensions` using interfaces.
- Updated tests to align with changes in validation and repository logic.

These changes promote a more modular, maintainable, and testable codebase.

+++ tests/ExampleLib.Tests/ValidationRunnerTests.cs
@@ -58,7 +58,7 @@ public class ValidationRunnerTests
     {
         var services = new ServiceCollection();
         services.AddSingleton<IApplicationNameProvider>(new StaticApplicationNameProvider("Tests"));
-        services.AddSaveValidation<YourEntity>(e => (decimal)e.Timestamp.Ticks, ThresholdType.RawDifference, 1m,
+        services.AddSaveValidation<YourEntity>(e => e.Timestamp.Ticks, ThresholdType.RawDifference, 1m,
             e => true);
         services.AddValidationRunner();
         var provider = services.BuildServiceProvider();
